refactor(commands): share cleanup plan resolver
