+++ src/commands/cleanup-plan.ts
@@ -1,0 +1,25 @@
+import type { OpenClawConfig } from "../config/config.js";
+import {
+  loadConfig,
+  resolveConfigPath,
+  resolveOAuthDir,
+  resolveStateDir,
+} from "../config/config.js";
+import { buildCleanupPlan } from "./cleanup-utils.js";
+
+export function resolveCleanupPlanFromDisk(): {
+  cfg: OpenClawConfig;
+  stateDir: string;
+  configPath: string;
+  oauthDir: string;
+  configInsideState: boolean;
+  oauthInsideState: boolean;
+  workspaceDirs: string[];
+} {
+  const cfg = loadConfig();
+  const stateDir = resolveStateDir();
+  const configPath = resolveConfigPath();
+  const oauthDir = resolveOAuthDir();
+  const plan = buildCleanupPlan({ cfg, stateDir, configPath, oauthDir });
+  return { cfg, stateDir, configPath, oauthDir, ...plan };
+}
